requirements.txt
streamlit
pandas
scikit-learn
numpy
requests
python-dotenv
matplotlib
